requirements.txt
pymupdf
openai
pinecone-client
python-dotenv
streamlit
tiktoken
watchdog